backend/services/vision_service.py
import cv2

def analyze_vision(video_path: str) -> dict:
    cap = cv2.VideoCapture(video_path)
    if not cap.isOpened():
        return {
            "eye_contact_score": 0.0,
            "smile_frequency": 0.0, 
            "head_movement_variance": 0.0,
            "posture_score": 0.0,
            "emotion_distribution": {"neutral": 1.0, "happy": 0.0, "nervous": 0.0}
        }
        
    fps = cap.get(cv2.CAP_PROP_FPS)
    if not fps or fps <= 0:
        fps = 30
        
    frame_interval = int(fps)
    
    frames_processed = 0
    faces_detected = 0
    
    # Use Haar Cascades as a robust, native fallback since MediaPipe solutions is broken on this version
    face_cascade = cv2.CascadeClassifier(cv2.data.haarcascades + 'haarcascade_frontalface_default.xml')
    
    while True:
        pos_frames = int(cap.get(cv2.CAP_PROP_POS_FRAMES))
        ret, frame = cap.read()
        if not ret:
            break
            
        if pos_frames % frame_interval == 0:
            frames_processed += 1
            gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
            # Detect faces
            faces = face_cascade.detectMultiScale(gray, scaleFactor=1.1, minNeighbors=5, minSize=(30, 30))
            if len(faces) > 0:
                faces_detected += 1
                
    cap.release()
    
    # Metrics calculation
    eye_contact_score = faces_detected / max(1, frames_processed)
    
    return {
        "eye_contact_score": round(eye_contact_score, 2),
        "smile_frequency": 0.6, # placeholder
        "head_movement_variance": 0.3, # placeholder
        "posture_score": 0.8, # placeholder
        "emotion_distribution": {
            "neutral": 0.6,
            "happy": 0.3,
            "nervous": 0.1
        }
    }
