backend/requirements.txt
fastapi~=0.104
uvicorn[standard]~=0.23
python-multipart~=0.0.6
transformers~=4.35
torch~=2.1
openai-whisper
opencv-python~=4.8
mediapipe~=0.10
numpy==2.4.4
moviepy~=1.0.3
pydantic~=2.4
scikit-learn~=1.3
